# tacocloud/tacocloud-api/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Taco Cloud Public API
  version: 1.0.0
  description: |
    Contrato oficial OpenAPI 3.0 para Taco Cloud (Laboratorio 6 - TC-35).
    Define los contratos de integración para ingredientes, diseño de tacos, pedidos con idempotencia
    y trazabilidad por Correlation ID, favoritos, cupones, pasarela de pago tokenizada, operaciones de cocina y
    anuncios operativos seguros. Incluye el modelo canónico de error RFC 7807 (ApiProblem).
  contact:
    name: Taco Cloud Engineering Team
    email: support@tacocloud.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desarrollo local (v1 Oficial)
  - url: http://localhost:8080/api
    description: Servidor de desarrollo local (Alias legado - Deprecado)
...
paths:
  /ingredients:
    get:
      tags: [Ingredients]
      summary: Obtener catálogo completo de ingredientes
      operationId: getAllIngredients
      responses:
        '200':
          description: Lista de ingredientes disponibles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/IngredientResponse'
    post:
      tags: [Ingredients]
      summary: Crear un nuevo ingrediente (Admin)
      operationId: createIngredient
      security:
        - basicAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngredientRequest'
      responses:
        '201':
          description: Ingrediente creado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
        '403':
          description: No autorizado (requiere rol ADMIN)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
  /ingredients/{id}:
    get:
      tags: [Ingredients]
      summary: Obtener un ingrediente por su identificador
      operationId: getIngredientById
      parameters:
        - name: id
          in: path
          required: true
          description: Código del ingrediente (ej. FLTO, CARN)
          schema:
            type: string
            example: FLTO
      responses:
        '200':
          description: Ingrediente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngredientResponse'
        '404':
          description: Ingrediente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiProblem'
    delete:
      tags: [Ingredients]
      summary: Eliminar un ingrediente (Admin)
      operationId: deleteIngredient
      security:
        - basicAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Ingrediente eliminado exitosamente
        '403':
          description: Requiere rol ADMIN
...
  /payment-methods/tokenize:
    post:
      tags: [Payments]
      summary: Tokenizar tarjeta de crédito para pagos futuros
      operationId: tokenizePaymentMethod
      security:
        - basicAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenizeRequest'
      responses:
        '201':
          description: Tarjeta tokenizada con éxito (CVV descartado de inmediato)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentMethodResponse'
        '400':
          description: Datos de tarjeta inválidos
  /payment-methods:
    get:
      tags: [Payments]
      summary: Listar métodos de pago guardados del usuario
      operationId: getPaymentMethods
      security:
        - basicAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Métodos de pago registrados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentMethodResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT para usuarios autenticados
    basicAuth:
      type: http
      scheme: basic
      description: Autenticación básica HTTP (usuario y contraseña)
...
  schemas:
...
    IngredientRequest:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          example: FLTO
        name:
          type: string
          example: Flour Tortilla
        type:
          type: string
          enum: [WRAP, PROTEIN, VEGGIES, CHEESE, SAUCE]
          example: WRAP
    IngredientResponse:
      type: object
      required:
        - id
        - name
        - type
      properties:
        id:
          type: string
          example: FLTO
        name:
          type: string
          example: Flour Tortilla
        type:
          type: string
          example: WRAP

    # --- TACOS ---
...
    TokenizeRequest:
      type: object
      required:
        - cardNumber
        - expiration
        - cvv
      properties:
        cardNumber:
          type: string
          example: 4111111111111111
        expiration:
          type: string
          example: 12/28
        cvv:
          type: string
          example: '123'
    PaymentMethodResponse:
      type: object
      properties:
        id:
          type: string
          example: pm_987654321
        brand:
          type: string
          example: Visa
        last4:
          type: string
          example: '1111'
        expiration:
          type: string
          example: 12/28
        maskedToken:
          type: string
          example: tok_***1111
        username:
          type: string
          example: habuma

    # --- COCINA ---
...
    ApiProblem:
      type: object
      required:
        - title
        - status
      properties:
        type:
          type: string
          format: uri
          example: about:blank
        title:
          type: string
          example: Bad Request
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: Request validation failed
        instance:
          type: string
          example: /api/v1/orders
        code:
          type: string
          example: VALIDATION_ERROR
        violations:
          type: array
          items:
            $ref: '#/components/schemas/ApiProblemViolation'
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
          example: 550e8400-e29b-41d4-a716-446655440000
    ApiProblemViolation:
      type: object
      properties:
        field:
          type: string
          example: deliveryName
        message:
          type: string
          example: Delivery name is required